+++ example/example.dart
@@ -8,13 +8,16 @@ import 'package:steel_crypt/steel_crypt.dart';
 
 main() {
 
-  var private = CryptKey().genKey(); //generate key
+  var private = CryptKey().genKey(); //generate 32 byte key
 
 
-  var encrypter = SymCrypt(private, 'AES'); //generate AES encrypter with key
+  var encrypter = AesCrypt(private, 'cbc'); //generate AES encrypter with key
 
 
   var encrypter2 = RsaCrypt(); //generate RSA encrypter
+
+
+  var encrypter3 = LightCrypt(private, "ChaCha20"); //generate ChaCha20 encrypter
 
 
   var hasher = HashCrypt(); //generate SHA-3/512 hasher
@@ -22,10 +25,13 @@ main() {
   var hasher2 = HashCrypt('SHA-3/256'); //generate SHA-3/256 hasher
 
 
-  var passHash = PassCrypt();
+  var passHash = PassCrypt(); //generate PBKDF2 password hasher
 
 
   var iv = CryptKey().genIV(16); //generate iv for AES
+
+  var iv2 = CryptKey().genIV(12); //generate iv for ChaCha20
+
 
   var salt = CryptKey().genIV(16); //generate salt for password hashing
 
@@ -43,9 +49,9 @@ main() {
 
   print(hasher.hash('word')); //perform hash
 
-  var hash = hasher.hash('word');
+  var hash = hasher.hash('pass');
 
-  print(hasher.checkhash('word', hash)); //perform check
+  print(hasher.checkhash('pass', hash)); //perform check
 
   print("");
 
@@ -74,7 +80,19 @@ main() {
   print("");
 
 
-  //AES with PKCS7 padding; Symmetric
+  //ChaCha20; Symmetric stream cipher
+  print("ChaCha20 Symmetric:");
+
+  print(encrypter3.encrypt('word', iv2)); //encrypt
+
+  String crypted3 = encrypter3.encrypt('word', iv2);
+
+  print(encrypter3.decrypt(crypted3, iv2)); //decrypt
+
+  print("");
+
+
+  //AES with PKCS7 padding; Symmetric block cipher
   print("AES Symmetric:");
 
   print(encrypter.encrypt('word', iv)); //encrypt